\name{astroC}
\alias{cJDJ2000}
\alias{cDAYPJULCENT}
\alias{cDAYPYEARTROP}
\alias{cDAYPYEARSID}
\alias{cDAYPMONSYN}
\alias{cDAYPMONSID}
\alias{cK}
\alias{cC}
\alias{cRE}
\alias{cMY}
\alias{cPRECESS}
\alias{cEPSOBL}
\alias{cAE}
\alias{cSBYE}
\alias{cSBYEMY}
\alias{cSBYME}
\alias{cSBYVE}
\alias{cSBYMA}
\alias{cSBYJU}
\alias{cSBYSA}
\alias{cSBYUR}
\alias{cSBYNE}
\alias{cSBYPL}
\alias{cSOLBYSID}
\alias{cSIDBYSOL}
\alias{DPY}
\alias{DAYINMONTH}
\title{ Astronomical constants}
\description{ Astronomical constants }
\details{
  \code{cJDJ2000} = 2451545.0 , Julian day number of the epoch J2000.0\cr
  \code{cDAYPJULCENT} = 36525.0 , days per julian century\cr
  \code{cDAYPYEARTROP} = 365.242198781 , days per tropical year\cr
  \code{cDAYPYEARSID} = 365.25636042 , days per sidereal year\cr
  \code{cDAYPMONSYN} = 29.53058868 , days per synodical month\cr
  \code{cDAYPMONSID} = 27.321655 , days per sidereal month\cr
  \code{cK} = 0.01720209895 , Gravitational constant, GAUSSian defintion\cr
  \code{cC} = 299792458.0  , [m/s] defined speed of light\cr
  \code{cRE} = 6378140.0  , [m]  radius of earth.s equator\cr
  \code{cMY} = 0.01230002  , ratio mass of moon/mass of earth\cr
  \code{cPRECESS} = 5029.0966  , [arc sec] precession per year at 2000.0\cr
  \code{cEPSOBL}23.43929111  , [deg]  inclination of ecliptic at 2000.0\cr  \code{cAE} = 1.49597870E11 , [m]  distance Earth to Sun\cr
  \code{cSBYE} = 332946.0   , ratio mass of Sun/mass of Earth\cr
  \code{cSBYEMY} = 328900.5  , ratio mass of Sun/mass of, Earth+Moon\cr
  \code{cSBYME} = 6023600.0  , ratio mass of Sun/mass of Mercury\cr
  \code{cSBYVE} = 408523.5   , ratio mass of Sun/mass of Venus\cr
  \code{cSBYMA} = 3098710.0  , ratio mass of Sun/mass of Mars\cr
  \code{cSBYJU} = 1047.355  , ratio mass of Sun/mass of Jupiter\cr
  \code{cSBYSA} = 3498.5  , ratio mass of Sun/mass of Saturn\cr
  \code{cSBYUR} = 22869.0  , ratio mass of Sun/mass of Uranus\cr
  \code{cSBYNE} = 19314.0  , ratio mass of Sun/mass of Neptun\cr
  \code{cSBYPL} = 130000000.0  , ratio mass of Sun/mass of Pluto\cr
  \code{cSOLBYSID} = 1.00273790934 , ratio solar/sidereal day\cr
  \code{cSIDBYSOL} = 0.99726956634 , ratio sidereal/solar day\cr
  \code{DPY} = cDAYPJULCENT/100.0; days/jul.Jahr\cr
  \code{DAYINMONTH} = c(31,28,31,30,31,30,31,31,30,31,30,31,31)
}
\author{  Christian W. Hoffmann <christian@echoffmann.ch>}
\keyword{arith}
