\name{Const}
\alias{c38}
\alias{c3Q}
\alias{GreatestIntAsRealF}
\alias{ASCII}
\alias{HexDig}
\alias{TAU}
\title{
Constants
}
\description{
Constants
}
\details{
  \code{c3Q} := .Machine$double.xmax^0.75\cr
  \code{c38} := sqrt(c3Q) \cr
  \code{ASCII} := ASCII characters corresponding to  (0), 1..256 \cr
  \code{HexDig} := '1' - '9', 'A' - 'F' \cr
  \code{TAU} := (1+sqrt(5))/2 # golden section constant = 1.6180
}
\author{  Christian W. Hoffmann <christian@echoffmann.ch>}
\keyword{arith}
