\name{cwhmisc-package}
\alias{cwhmisc-package}
\title{A collection of useful functions and constants}
\description{
Useful functions and constants for
 mathematics, astronomy, plotting, printing, data manipulation, statistics, string manipulation, etc.
}
\details{
\tabular{ll}{
Package: \tab cwhmisc\cr
Type: \tab Package\cr
Version: \tab 5.0\cr
Date: \tab 2014-08-20\cr
License: \tab GPL (>= 2)\cr
Author: \tab  Christian W. Hoffmann <christian@echoffmann.ch>\cr
Maintainer: \tab same
}
}
