\name{invgauss}
\alias{dinvgauss}
\alias{pinvgauss}
\alias{qinvgauss}
\alias{rinvgauss}
\title{Inverse Gaussian Distribution}
\description{
  Density, cumulative probability, quantiles and random generation for the inverse Gaussian distribution.
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au} }
\keyword{distribution}
