\name{libs}
\alias{libs}
\title{List all installed packages, or all functions in a package}
\description{
  Lists all packages (called without an argument) or
the functions in a package (called with the package name - quotes not
needed).
}
\usage{
  libs(Lib)
}
\arguments{
  \item{Lib}{package name, if missing see above}
}
\examples{
  libs()
  libs(base)
}
\author{
??
}

\keyword{manip}
