\name{lpr}
\alias{lpr}
\title{Print an object}
\description{Print a given object}
\usage{
lpr(object, file="Rplotlpr.ps", ...)
}
\arguments{
  \item{object}{The object to be printed. If missing, the current plot will be printed.}
  \item{file}{file to receive printed version.}
  \item{\dots}{Additional parameters for \code{\link{dev.copy}}.}
}
\author{Ray Brownrigg \email{ray@mcs.vuw.ac.nz} \cr
modified by Christian W. Hoffmann  \email{christian@echoffmann.ch}
}
\keyword{device}
