\name{parsecheck}
\alias{parsecheck}
\title{check files for parsing errors}
\description{
  check files for parsing errors
}
\usage{
parsecheck(str)
}
\arguments{
  \item{str}{Directory containing *.R files to examine}
}
\value{
  file name and place where parsing error occurred; mostly missing brackets/braces
}
\author{
  Duncan Murdoch via  "unable to collate and parse R files"
}
\keyword{misc}
\keyword{language}
