\name{poisgam}
\alias{dpoisgam}
\alias{ppoisgam}
\alias{qpoisgam}
\alias{ppoiscc}
\alias{rpoisgam}
\title{Poisson Gamma Distribution}
\description{
  Density, cumulative probability function and random variates for the Poisson-gamma or
     compound Poisson distribution. The Poison-gamma distribution is a Tweedie distribution
     with index p between 1 and 2.  

	See: 	\url{http://www.statsci.org/smyth/index.html}
}
\author{
Gordon Smyth, \email{gks@maths.uq.edu.au}
}

\keyword{distribution}
