\name{qres}
\alias{qres.binom}
\alias{qres.gamma}
\alias{qres.invgauss}
\alias{qres.nbinom}
\alias{qres.pois}
\title{Randomized quantile residuals}
\description{
  Computes randomized quantile residuals for binomial, Poisson, negative
	binomial, gamma and inverse Gaussian generalized linear models.

	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/qres.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au}}
\seealso{
  \itemize{
  \item \url{http://www.maths.uq.edu.au/~gks/s/tweedie.html} Tweedie Distributions,  \item \url{http://www.maths.uq.edu.au/~gks/s/tweedief.html} Tweedie family,
  \item \url{http://www.maths.uq.edu.au/~gks/s/poisgam.html} Poison-gamma Distribution,
  \item \url{http://www.maths.uq.edu.au/~gks/s/invgauss.html} Inverse Gaussian Distribution.
  }
}
\keyword{distribution}
