\name{setPPT}
\alias{setPPT}
\title{Set PowerPoint style}
\description{
  Set par() with suitable options for Power point slides.
	Note that you have to set them every time you open up a new device.
	Then save the pictures to PNG format with 'width=5' and 'height=5' pointsize='14'.
}
\usage{
setPPT()
}
\author{
Henrik Bengtsson, \email{hb@maths.lth.se}
}
\keyword{documentation}

