\name{tweedie}
\alias{tweedie}  # why is this necessary?
\alias{dtweedie}
\alias{ptweedie}
\alias{qtweedie}
\alias{saddle.tweedie}
\alias{devi.tweedie}
\alias{tail.tweedie}
\alias{tweedief}
\title{Tweedie Distributions}
\description{
  Density, cumulative distribution function and quantiles for the
	Tweedie distributions. Includes the normal, Poison, Poison-gamma and inverse-Gaussian distributions as special cases. 

tweedief  produces a generalized linear model family object with any power
	variance function and any power link. Includes the Gaussian, Poisson,
	gamma and inverse-Gaussian families as special cases.
	See:	\url{http://www.statsci.org/smyth/index.html}
}
\author{
Gordon Smyth, \email{gks@maths.uq.edu.au},
    with reservations included by:
  Christian W. Hoffmann, \email{c-w.hoffmann@sunrise.ch}, \cr
  \url{http://www.wsl.ch/personal_homepages/hoffmann/index_EN?-C=&n} \cr
}
\keyword{distribution}
