\name{delstr}
\alias{delstr}
\title{Delete a substring from a string}
\description{
  \code{delstr} deletes a substring from a string
}
\usage{
delstr(str,del)
}
\arguments{
  \item{str} {a string, may be empty.}
  \item{del} {a string, may be empty.}
}
\value{
 A string
}
\examples{
delstr("Ich gehen heuten in den Garten hinein","en")
#  "Ich geh heut in d Gart hinein"}
\author{
  Christian W. Hoffmann <c-w.hoffmann@sunrise.ch> \cr
  \url{http://www.wsl.ch/personal_homepages/hoffmann/index_EN}
}
\keyword{print}
